requirements.txt
asgiref==3.2.10
autopep8==1.5.4
cffi==1.14.2
cryptography==3.1
Django==3.1
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
waitress==1.4.4
